+++ lib/restclient/payload.rb
@@ -149,7 +149,9 @@ module RestClient
 
       def create_file_field(s, k, v)
         begin
-          s.write("Content-Disposition: form-data; name=\"#{k}\"; filename=\"#{v.respond_to?(:original_filename) ? v.original_filename : File.basename(v.path)}\"#{EOL}")
+          s.write("Content-Disposition: form-data;")
+          s.write(" name=\"#{k}\";") unless (k.nil? || k=='')
+          s.write(" filename=\"#{v.respond_to?(:original_filename) ? v.original_filename : File.basename(v.path)}\"#{EOL}")
           s.write("Content-Type: #{v.respond_to?(:content_type) ? v.content_type : mime_for(v.path)}#{EOL}")
           s.write(EOL)
           while data = v.read(8124)