+++ history.md
@@ -1,6 +1,7 @@
-# 1.4.3
+# 1.5.0
 
-- added Response.to_str and AbstractResponse.to_i to improve semantic and compatibility
+- the response is now a string with the Response module
+- added AbstractResponse.to_i to improve semantic
 - multipart Payloads ignores the name attribute if it's not set (patch provided by Tekin Suleyman)
 - correctly takes into account user headers whose keys are strings (path provided by Cyril Rohr)